+++ src/main/java/versus/repository/interfaces/TeamRepository.java
@@ -5,9 +5,9 @@ import versus.model.Team;
 import java.util.List;
 
 public interface TeamRepository {
-    Team getTeamById(int id);
+    Team getTeamById(long id);
     List<Team> getAllTeams();
     Team createTeam(Team team);
     Team updateTeam(Team team);
-    void deleteTeam(int id);
+    void deleteTeam(long id);
 }
